Revert "Web"

+++ src/main/java/ru/coolteam/earnpocketmoney/repositories/ParentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.coolteam.earnpocketmoney.models.Parent;
 
 @Repository
-public interface ParentRepository extends JpaRepository<Parent, Integer> {
+public interface ParentRepository extends JpaRepository<Parent, Long> {
 }
